perf(resp): split TCP without BiLock and shrink ArgLayout to 12 bytes
Plain-TCP connections now use `TcpStream::into_split()` (owned halves) instead
of `tokio::io::split`, dropping the per-read/write BiLock. TLS and async-std
keep `io::split`. Separately, `ArgLayout.start` moves from u64 to u32, shrinking
the struct from 16 to 12 bytes; command payloads never exceed u32 range.

+++ src/resp/arg_serializer.rs
@@ -69,7 +69,7 @@ impl<'a> ArgSerializer<'a> {
         // 4. Update the layout index
         if let Some(ref mut layout) = self.args_layout {
             layout.push(ArgLayout {
-                start: start_pos as u64,
+                start: start_pos as u32,
                 len: data_len as u32,
                 slot: 0,
                 flags: 0,
